# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Transferências e Usuários",
    "version": "1.0.0",
    "description": "API para registro, login, consulta de usuários e transferências."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
    "/register": {
      "post": {
        "summary": "Registrar novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "favorecidos": { "type": "array", "items": { "type": "string" } }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário criado" },
          "409": { "description": "Usuário já existe" }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Login de usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login bem-sucedido" },
          "401": { "description": "Credenciais inválidas" }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Listar usuários",
        "responses": {
          "200": { "description": "Lista de usuários" }
        }
      }
    },
    "/users/{username}/favorecidos": {
      "post": {
        "summary": "Adicionar favorecido",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "favorecido": { "type": "string" }
                },
                "required": ["favorecido"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Favorecido adicionado" },
          "404": { "description": "Usuário não encontrado" }
        }
      }
    },
    "/transfer": {
      "post": {
        "summary": "Realizar transferência",
        "security": [ { "bearerAuth": [] } ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "remetente": { "type": "string" },
                  "destinatario": { "type": "string" },
                  "valor": { "type": "number" }
                },
                "required": ["remetente", "destinatario", "valor"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Transferência realizada" },
          "400": { "description": "Erro na transferência" },
          "401": { "description": "Token não fornecido" },
          "403": { "description": "Token inválido" }
        }
      }
    },
    "/transfers": {
      "get": {
        "summary": "Listar transferências",
        "security": [ { "bearerAuth": [] } ],
        "responses": {
          "200": { "description": "Lista de transferências" },
          "401": { "description": "Token não fornecido" },
          "403": { "description": "Token inválido" }
        }
      }
    }
  }
}